// src/models/TextMetadata.ts
export interface TextMetadata {
	fontFamily?: string;
	fontSizePx?: number;
	padding?: number;
	vpadding?: number;
	spacing?: number;
	fgColour?: string;
	bgColour?: string;
	align?: string;
	borderColour?: string;
	borderWidth?: string;
	borderDash?: Array<number>;
	bold?: boolean;
	italic?: boolean;
}